refactor: rename func-call-spacing to function-call-spacing
