+++ packages/eslint-plugin-js/rules.md
@@ -15,7 +15,7 @@
 | [`@stylistic/js/computed-property-spacing`](./rules/computed-property-spacing) | Enforce consistent spacing inside computed property brackets | ✅ |  |
 | [`@stylistic/js/dot-location`](./rules/dot-location) | Enforce consistent newlines before and after dots | ✅ |  |
 | [`@stylistic/js/eol-last`](./rules/eol-last) | Require or disallow newline at the end of files | ✅ |  |
-| [`@stylistic/js/func-call-spacing`](./rules/func-call-spacing) | Require or disallow spacing between function identifiers and their invocations | ✅ |  |
+| [`@stylistic/js/function-call-spacing`](./rules/function-call-spacing) | Require or disallow spacing between function identifiers and their invocations | ✅ |  |
 | [`@stylistic/js/function-call-argument-newline`](./rules/function-call-argument-newline) | Enforce line breaks between arguments of a function call | ✅ |  |
 | [`@stylistic/js/function-paren-newline`](./rules/function-paren-newline) | Enforce consistent line breaks inside function parentheses | ✅ |  |
 | [`@stylistic/js/generator-star-spacing`](./rules/generator-star-spacing) | Enforce consistent spacing around `*` operators in generator functions | ✅ |  |
