+++ packages/eslint-plugin-js/rules/function-call-spacing/function-call-spacing.js
@@ -23,7 +23,7 @@ module.exports = {
     docs: {
       description: 'Require or disallow spacing between function identifiers and their invocations',
       recommended: false,
-      url: 'https://eslint.style/rules/js/func-call-spacing',
+      url: 'https://eslint.style/rules/js/function-call-spacing',
     },
 
     fixable: 'whitespace',
