+++ packages/eslint-plugin-js/rules/function-call-spacing/types.d.ts
