+++ packages/eslint-plugin-js/rules/index.js
@@ -12,7 +12,7 @@ module.exports = {
   'computed-property-spacing': require('./computed-property-spacing/computed-property-spacing.js'),
   'dot-location': require('./dot-location/dot-location.js'),
   'eol-last': require('./eol-last/eol-last.js'),
-  'func-call-spacing': require('./func-call-spacing/func-call-spacing.js'),
+  'function-call-spacing': require('./function-call-spacing/function-call-spacing.js'),
   'function-call-argument-newline': require('./function-call-argument-newline/function-call-argument-newline.js'),
   'function-paren-newline': require('./function-paren-newline/function-paren-newline.js'),
   'generator-star-spacing': require('./generator-star-spacing/generator-star-spacing.js'),
