+++ packages/eslint-plugin-js/rules/keyword-spacing/keyword-spacing.test.js
@@ -1074,7 +1074,7 @@ ruleTester.run('keyword-spacing', rule, {
     { code: 'class A extends B { constructor() { ({a:super() }) } }', parserOptions: { ecmaVersion: 6 } },
     { code: 'class A extends B { constructor() { ({a: super() }) } }', options: [NEITHER], parserOptions: { ecmaVersion: 6 } },
 
-    // not conflict with `func-call-spacing`
+    // not conflict with `function-call-spacing`
     { code: 'class A extends B { constructor() { super(); } }', parserOptions: { ecmaVersion: 6 } },
     { code: 'class A extends B { constructor() { super (); } }', options: [NEITHER], parserOptions: { ecmaVersion: 6 } },
