+++ packages/eslint-plugin-ts/rules/function-call-spacing/README.md
@@ -3,5 +3,5 @@ description: 'Require or disallow spacing between function identifiers and their
 ---
 
 
-This rule extends the base [`func-call-spacing`](/rules/js/func-call-spacing) rule.
+This rule extends the base [`function-call-spacing`](/rules/js/function-call-spacing) rule.
 It adds support for generic type parameters on function calls.
