+++ packages/eslint-plugin-ts/rules/function-call-spacing/types.d.ts
