+++ packages/metadata/src/metadata.ts
@@ -198,11 +198,11 @@ export const packages: Readonly<PackageInfo[]> = Object.freeze([
         }
       },
       {
-        "name": "func-call-spacing",
-        "ruleId": "@stylistic/js/func-call-spacing",
-        "originalId": "func-call-spacing",
-        "entry": "packages/eslint-plugin-js/rules/func-call-spacing/func-call-spacing.js",
-        "docsEntry": "packages/eslint-plugin-js/rules/func-call-spacing/README.md",
+        "name": "function-call-spacing",
+        "ruleId": "@stylistic/js/function-call-spacing",
+        "originalId": "function-call-spacing",
+        "entry": "packages/eslint-plugin-js/rules/function-call-spacing/function-call-spacing.js",
+        "docsEntry": "packages/eslint-plugin-js/rules/function-call-spacing/README.md",
         "meta": {
           "fixable": "whitespace",
           "docs": {
@@ -1276,11 +1276,11 @@ export const packages: Readonly<PackageInfo[]> = Object.freeze([
         }
       },
       {
-        "name": "func-call-spacing",
-        "ruleId": "@stylistic/ts/func-call-spacing",
-        "originalId": "@typescript-eslint/func-call-spacing",
-        "entry": "packages/eslint-plugin-ts/rules/func-call-spacing/func-call-spacing.ts",
-        "docsEntry": "packages/eslint-plugin-ts/rules/func-call-spacing/README.md",
+        "name": "function-call-spacing",
+        "ruleId": "@stylistic/ts/function-call-spacing",
+        "originalId": "@typescript-eslint/function-call-spacing",
+        "entry": "packages/eslint-plugin-ts/rules/function-call-spacing/function-call-spacing.ts",
+        "docsEntry": "packages/eslint-plugin-ts/rules/function-call-spacing/README.md",
         "meta": {
           "fixable": "whitespace",
           "docs": {
@@ -1674,10 +1674,10 @@ export const packages: Readonly<PackageInfo[]> = Object.freeze([
         }
       },
       {
-        "name": "func-call-spacing",
-        "ruleId": "@stylistic/func-call-spacing",
-        "entry": "packages/eslint-plugin-js/rules/func-call-spacing/func-call-spacing.js",
-        "docsEntry": "packages/eslint-plugin-js/rules/func-call-spacing/README.md",
+        "name": "functon-call-spacing",
+        "ruleId": "@stylistic/function-call-spacing",
+        "entry": "packages/eslint-plugin-js/rules/function-call-spacing/function-call-spacing.js",
+        "docsEntry": "packages/eslint-plugin-js/rules/function-call-spacing/README.md",
         "meta": {
           "fixable": "whitespace",
           "docs": {
